Changes to EDF-Viewer: (1) The reason to those changes is that we want to integrate Matlab-based EDF-Viewer as an external module of Java-based EDF-Editor. Current changes include A) allowing invoking EDF-Viewer from the common line; B) visualizing pre-specified EDF and annotation files when the Viewer starts up;
(2) Change to "EDF_View.m" includes:
#1 We allowed EDF-Viewer to accept parameters that users want to
visualize a EDF with its corresponding annotation file (if any) without
interacting with GUI.
#2 We fixed a bug that cause reading or updating "dirty data" within
multiple functions.

(3) Create "StartMe_cmdPars.m"
This is one of the two new entry points (instead of EDF_View.m) we made
to EDF-Viewer. This allows EDF-Viewer to be invokded from command line
with parameters a user specified.  The parameters include filename and
filepath of a EDF file, and those of an annotation file (if any).  With
those parameters, the EDF-Viewer will start up and initially visualize
the chosen EDF and annotation files.  Users can switch to other EDF
files via dialog selection as well.

(4) Create "StartMe_dialogPars.m"
This is one of the two new entry points (instead of EDF_View.m) we made
to EDF-Viewer. This allows EDF-Viewer behaving like the original
EDF_View.m, that is it can start up with en empty selection and users
need to select EDF and annotation files via dialog selection.